+++ src/integrationTest/java/uk/gov/hmcts/opal/AbstractIntegrationTest.java
@@ -1,0 +1,33 @@
+package uk.gov.hmcts.opal;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static uk.gov.hmcts.opal.TestContainerConfig.POSTGRES_CONTAINER;
+
+@SpringBootTest
+@ActiveProfiles("integration")
+@ContextConfiguration(classes = {TestContainerConfig.class})
+@AutoConfigureMockMvc
+public class AbstractIntegrationTest {
+    @Autowired
+    protected MockMvc mockMvc;
+
+    @Autowired
+    protected ObjectMapper objectMapper;
+
+    // Dynamically register properties to configure the datasource
+    @DynamicPropertySource
+    static void databaseProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url", POSTGRES_CONTAINER::getJdbcUrl);
+        registry.add("spring.datasource.username", POSTGRES_CONTAINER::getUsername);
+        registry.add("spring.datasource.password", POSTGRES_CONTAINER::getPassword);
+    }
+}
