+++ src/integrationTest/java/uk/gov/hmcts/opal/TestContainerConfig.java
@@ -1,0 +1,20 @@
+package uk.gov.hmcts.opal;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.utility.DockerImageName;
+
+@TestConfiguration
+public class TestContainerConfig {
+
+    public static final PostgreSQLContainer<?> POSTGRES_CONTAINER;
+
+    static {
+        POSTGRES_CONTAINER = new PostgreSQLContainer<>(DockerImageName.parse("postgres:17"))
+            .withDatabaseName("testdb")
+            .withUsername("test")
+            .withPassword("test");
+
+        POSTGRES_CONTAINER.start();
+    }
+}
