+++ src/integrationTest/java/uk/gov/hmcts/opal/database/FlywayIntegrationTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import uk.gov.hmcts.opal.BaseIntegrationTest;
 
 import java.util.Arrays;
 import java.util.List;
@@ -15,7 +14,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @SpringBootTest
 @ActiveProfiles("integration")
-public class FlywayIntegrationTest extends BaseIntegrationTest {
+public class FlywayIntegrationTest {
 
     @Autowired
     Flyway flyway;
