+++ src/integrationTest/java/uk/gov/hmcts/opal/openapi/OpenAPIPublisherTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.web.servlet.MockMvc;
-import uk.gov.hmcts.opal.BaseIntegrationTest;
 
 import java.io.OutputStream;
 import java.nio.file.Files;
@@ -19,7 +19,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  * Each CI run on master should automatically save and upload (if updated) documentation.
  */
 @AutoConfigureMockMvc
-class OpenAPIPublisherTest extends BaseIntegrationTest {
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class OpenAPIPublisherTest {
 
     @Autowired
     private MockMvc mvc;
